+++ app/Repository/TalkRepository.php
@@ -81,13 +81,13 @@ class TalkRepository extends ServiceEntityRepository
         $queryBuilder
             ->select('COUNT(o.id) AS total')
             ->andWhere(
-                $queryBuilder->expr()->gte('o.startsAt', ':start_date'),
+                $queryBuilder->expr()->gte('o.startsAt', ':startDate'),
             )
             ->andWhere(
-                $queryBuilder->expr()->lt('o.startsAt', ':end_date'),
+                $queryBuilder->expr()->lt('o.startsAt', ':endDate'),
             )
-            ->setParameter('start_date', $datePeriod->getStartDate())
-            ->setParameter('end_date', $datePeriod->getEndDate())
+            ->setParameter('startDate', $datePeriod->getStartDate())
+            ->setParameter('endDate', $datePeriod->getEndDate())
         ;
 
         foreach ($groupBy as $name => $select) {
