+++ config/packages/doctrine.yaml
@@ -24,9 +24,9 @@ doctrine:
                 alias: App
         dql:
             string_functions:
-                DAY: 'App\Doctrine\Day'
-                MONTH: 'App\Doctrine\Month'
-                YEAR: 'App\Doctrine\Year'
+                DAY: 'App\Doctrine\Query\AST\Function\Day'
+                MONTH: 'App\Doctrine\Query\AST\Function\Month'
+                YEAR: 'App\Doctrine\Query\AST\Function\Year'
 
 when@test:
     doctrine:
